// components/Interfaces/UserData.ts
export interface UserData {
  address?: string;
  business_name?: string;
  city?: string;
  created_at?: string;
  ein_number?: string;
  email?: string;
  id?: number;
  image_data?: string;
  is_active?: boolean;
  is_staff?: boolean;
  last_login?: string;
  password?: string;
  phone_number?: string;
  role?: string;
  state?: string;
  updated_at?: string;
  username?: string;
  zipcode?: string;
}